Fix: login.js, signup.js and add module.js
- 로그인, 회원가입 페이지에 공통으로 사용하는 공통된 로직을 모듈로 분리.
