+++ login/login.css
@@ -116,14 +116,14 @@ label input:focus {
   border-color: var(--blue-color);
 }
 #password{
-  margin-bottom: 3rem;
+  margin-bottom: 0.5rem;
 }
 #eyeImg {
   width: 1.5rem;
   height: 1.5rem;
   position: relative;
   left: 36.875rem;
-  bottom: 6.3rem;
+  bottom: 3.9rem;
 }
 
 form button {
