+++ login/login.html
@@ -66,6 +66,6 @@
     <span>판다마켓이 처음이신가요?</span>
     <a href="../signup/signup.html">회원가입</a>
   </footer>
-  <script src="login.js"></script>
+  <script src="login.js" type="module"></script>
 </body>
 </html>
