+++ login/login.js
@@ -1,27 +1,13 @@
-const USER_DATA = [
-  { email: 'codeit1@codeit.com', password: "codeit101!" },
-  { email: 'codeit2@codeit.com', password: "codeit202!" },
-  { email: 'codeit3@codeit.com', password: "codeit303!" },
-  { email: 'codeit4@codeit.com', password: "codeit404!" },
-  { email: 'codeit5@codeit.com', password: "codeit505!" },
-  { email: 'codeit6@codeit.com', password: "codeit606!" },
-];
+import {USER_DATA, formEl, labelEl, emailEl, passwordEl, passwordToggle,
+  focusoutEmail
+} from '../module.js';
 
 /* getting element */
-const formEl = document.querySelector('form');
-const labelEl = document.querySelectorAll('label')
-const emailLabelEl = labelEl[0];
 const passwordLabelEl = labelEl[1];
-const emailEl = document.querySelector('#email');
-const passwordEl = document.querySelector('#password');
 const btnEl = document.querySelector('#loginBtn');
 const eyeImgEl = document.querySelector('#eyeImg');
 
 /* setting variable and fixing html  */
-const emailNewP = document.createElement('p');
-emailNewP.classList.add("errorMsg");
-emailLabelEl.append(emailNewP);
-
 const passwordNewP = document.createElement('p');
 passwordNewP.classList.add('errorMsg');
 passwordNewP.classList.add('passwordBottom')
@@ -30,28 +16,7 @@ passwordLabelEl.append(passwordNewP);
 btnEl.setAttribute('disabled', true);
 
 /* event handler */
-function validateEmail(email) {
-  const emailPattern = /^[^\s@]+@[^\s@]+\.(com)$/;
-  return emailPattern.test(email);
-}
 
-function focusoutEmail() {
-  if(emailEl.value === ""){ 
-    emailEl.classList.add('borderRed');
-    emailNewP.textContent = "이메일을 입력해주세요.";
-    emailNewP.style.display = 'block';
-    return false;
-  } else if(!validateEmail(emailEl.value)) {
-    emailEl.classList.add('borderRed');
-    emailNewP.textContent = "잘못된 이메일 형식입니다.";
-    emailNewP.style.display = 'block';
-    return false;
-  } else {
-    emailEl.classList.remove('borderRed');
-    emailNewP.style.display = 'none';
-    return true;
-  }
-}
 
 function focusoutPassword() {
   if(passwordEl.value === "") {
@@ -98,15 +63,6 @@ function conditionalBtn() {
     btnEl.addEventListener('click', clickLink);
   } 
 }
-
-function passwordToggle() {
-  if(passwordEl.type === 'password'){
-    passwordEl.setAttribute('type', 'text');
-  } else{
-    passwordEl.setAttribute('type', 'password');
-  }
-}
-
 /* event handling */ 
 emailEl.addEventListener('focusout', focusoutEmail);
 passwordEl.addEventListener('focusout', focusoutPassword);
