+++ signup/signup.html
@@ -75,6 +75,6 @@
     <span>이미 회원이신가요?</span>
     <a href="../login/login.html">로그인</a>
   </footer>
-  <script src="signup.js"></script>
+  <script src="signup.js" type="module"></script>
 </body>
 </html>
