+++ signup/signup.js
@@ -1,21 +1,11 @@
-const USER_DATA = [
-  { email: 'codeit1@codeit.com', password: "codeit101!" },
-  { email: 'codeit2@codeit.com', password: "codeit202!" },
-  { email: 'codeit3@codeit.com', password: "codeit303!" },
-  { email: 'codeit4@codeit.com', password: "codeit404!" },
-  { email: 'codeit5@codeit.com', password: "codeit505!" },
-  { email: 'codeit6@codeit.com', password: "codeit606!" },
-];
+import {USER_DATA, formEl, labelEl, emailEl, passwordEl, emailNewP,
+  passwordToggle, focusoutEmail
+} from '../module.js';
 
-/* getting element */
-const formEl = document.querySelector('form');
-const labelEl = document.querySelectorAll('label')
-const emailLabelEl = labelEl[0];
+/* getting element */;
 const nicknameLabelEl = labelEl[1];
 const passwordLabelEl = labelEl[2];
 const passwordCheckLabelEl = labelEl[3];
-const emailEl = document.querySelector('#email');
-const passwordEl = document.querySelector('#password');
 const nicknameEl = document.querySelector('#nickname');
 const passwordCheckEl = document.querySelector('#passwordCheck');
 const btnEl = document.querySelector('#signUpBtn');
@@ -23,10 +13,6 @@ const eyeImgFirstEl = document.querySelector('.eyeImg');
 const eyeImgSecondEl = document.querySelector('.checkImg');
 
 /* setting variable and fixing html  */
-const emailNewP = document.createElement('p');
-emailNewP.classList.add("errorMsg");
-emailLabelEl.append(emailNewP);
-  
 const passwordNewP = document.createElement('p');
 passwordNewP.classList.add('errorMsg');
 passwordNewP.classList.add('passwordBottom')
@@ -44,29 +30,6 @@ nicknameLabelEl.append(nicknameNewP);
 btnEl.setAttribute('disabled', true);
 
 /* event handler */
-function validateEmail(email) {
-  const emailPattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  return emailPattern.test(email);
-}
-
-function focusoutEmail() {
-  if(emailEl.value === ""){ 
-    emailEl.classList.add('borderRed');
-    emailNewP.textContent = "이메일을 입력해주세요.";
-    emailNewP.style.display = 'block';
-    return false;
-  } else if(!validateEmail(emailEl.value)) {
-    emailEl.classList.add('borderRed');
-    emailNewP.textContent = "잘못된 이메일 형식입니다.";
-    emailNewP.style.display = 'block';
-    return false;
-  } else {
-    emailEl.classList.remove('borderRed');
-    emailNewP.style.display = 'none';
-    return true;
-  }
-}
-
 function focusoutPassword() {
   if(passwordEl.value === "") {
     passwordEl.classList.add('borderRed');
@@ -135,14 +98,6 @@ function conditionalBtn(){
     alert('사용 중인 이메일입니다.')
   } else {
     btnEl.addEventListener('click', loginLink);
-  }
-}
-
-function passwordToggle() {
-  if(passwordEl.type === 'password'){
-    passwordEl.setAttribute('type', 'text');
-  } else{
-    passwordEl.setAttribute('type', 'password');
   }
 }
